+++ app/seeds/letter_boards.py
@@ -1,11 +1,24 @@
-from app.models import Letter_Board
+from app.api.utils import find_all_words, find_all_rotations
+from app.models import Letter_Board, db
 
-def seed_users():
-  letter_board1 = Letter_Board(letters=['ansm', 'eall', 'vezs', 'icra'],
-                              words=['arcana', 'arsenic', 'earl', 'nerve', 'scare', 'miss', 'arrival', 'evince', 'airmass', 'incise', 'service', 'assassin', 'clear', 'vii', 'man', 'van', 'ran', 'alma', 'sesame', 'lascar', 'leave', 'minima', 'sis', 'sale', 'malaise', 'smile', 'simmer', 'sneeze', 'calla', 'lair', 'enamel', 'arisen', 'simile', 'cleanse', 'crime', 'raisin', 'airmen', 'ravel', 'scenic', 'cameraman', 'ami', 'lace', 'animal', 'arsenal', 'vaccine', 'serine', 'miasma', 'venereal', 'acme', 'mailmen', 'revel', 'seminar', 'enclave', 'arm', 'levee', 'mailman', 'arrear', 'clam', 'emissive', 'crania', 'massacre', 'mescaline', 'nice', 'crevice', 'lemma', 'servicemen', 'annal', 'mercer', 'slain', 'aerial', 'mare', 'vassal', 'mane', 'miasmal', 'mere', 'sneer', 'asinine', 'lava', 'senile', 'vile', 'assail', 'era', 'slam', 'silicic', 'cleric', 'ermine', 'nil', 'sec', 'seen', 'ascii', 'minim', 'viscera', 'sera', 'nave', 'sisal', 'slim', 'serene', 'vernier', 'verse', 'averse', 'vine', 'ram', 'evasive', 'camel', 'cancel', 'icicle', 'clavicle', 'river', 'amra', 'villain', 'mall', 'serviceman', 'sine', 'elves', 'crass', 'cannel', 'mesmeric', 'salesman', 'lim', 'seismic', 'cairn', 'reverie', 'canvas', 'maser', 'essence', 'caramel',
-                                    'niacin', 'silver', 'classic', 'sac', 'acre', 'raise', 'vermeil', 'cave', 'elm', 'viva', 'clinician', 'circa', 'access', 'vein', 'aile', 'laissez', 'mice', 'vanilla', 'rival', 'carve', 'sizzle', 'snell', 'niece', 'ell', 'risen', 'zeal', 'cane', 'rave', 'snivel', 'even', 'cerise', 'vicinal', 'same', 'scrim', 'siva', 'ceil', 'alliance', 'arcsine', 'several', 'sieve', 'arena', 'lineman', 'level', 'azalea', 'melanism', 'renaissance', 'calm', 'mar', 'camilla', 'screen', 'rime', 'savanna', 'rev', 'cameramen', 'small', 'rim', 'mill', 'melee', 'rinse', 'serial', 'science', 'since', 'malaria', 'lam', 'reave', 'veneer', 'via', 'velar', 'isle', 'never', 'rear', 'scriven', 'mezzanine', 'area', 'arcsin', 'acacia', 'ease', 'mean', 'mesa', 'mila', 'sans', 'crease', 'ill', 'nee', 'sire', 'crane', 'maim', 'ana', 'amnesiac', 'name', 'reminisce', 'maria', 'carven', 'series', 'sin', 'scarves', 'vernal', 'cancer', 'rascal', 'vale', 'mine', 'salami', 'caiman', 'carne', 'malarial', 'lilac', 'immense', 'revive', 'eel', 'casein', 'sell', 'aria', 'vase', 'livre', 'sincere', 'vial', 'vicar', 'raven', 'lama', 'valve', 'craze', 'zinnia', 'viii', 'leer', 'airmail', 'aniline', 'naval', 'mammalian', 'rill', 'allele', 'veer', 'canal', 'criss', 'increase', 'minimal', 'ice', 'min', 'cam', 'nail',
-                                    'scarce', 'acclaim', 'nirvana', 'sir', 'line', 'lien', 'simian', 'anneal', 'lessen', 'menial', 'maze', 'scream', 'lass', 'rise', 'eave', 'caveman', 'alien', 'mammal', 'revere', 'circle', 'evil', 'inimical', 'relieve', 'cavalier', 'ceramic', 'siren', 'ala', 'mana', 'avian', 'can', 'lac', 'scan', 'smear', 'arrive', 'reversal', 'riven', 'craven', 'manic', 'race', 'iii', 'male', 'nine', 'anise', 'rail', 'saliva', 'lance', 'lessee', 'mascara', 'zinc', 'clean', 'slice', 'linemen', 'caresses', 'veil', 'smell', 'scene', 'erase', 'cleave', 'vivace', 'villein', 'seance', 'armillaria', 'merlin', 'arise', 'alizarin', 'amen', 'incline', 'amnesia', 'versa', 'ani', 'sever', 'civilian', 'clinic', 'rare', 'lea', 'sarcasm', 'emma', 'relic', 'ale', 'cram', 'selves', 'slime', 'cavemen', 'menace', 'anvil', 'rein', 'lee', 'seem', 'all', 'snail', 'incense', 'silica', 'alarm', 'canine', 'cereal', 'millenia', 'revival', 'elan', 'mescal', 'sail', 'aver', 'vermin', 'careen', 'rice', 'cilia', 'err', 'sleeve', 'lame', 'sen', 'calve', 'insane', 'civil', 'marine', 'mace', 'amass', 'lime', 'near', 'seminarian', 'measle', 'arc', 'llama', 'valine', 'inane', 'arcane', 'carmine', 'naive', 'career', 'vane', 'animism', 'missive', 'sienna', 'sense', 'real', 'racial', 'arsine', 'vee', 'mail', 'scalar',
-                                    'viral', 'visceral', 'sima', 'melanin', 'navel', 'clime', 'variac', 'rever', 'iceman', 'semi', 'inverse', 'class', 'ire', 'carnival', 'crises', 'earn', 'mainline', 'sea', 'vessel', 'mimic', 'seize', 'camera', 'lens', 'mineral', 'sial', 'ravine', 'avarice', 'laze', 'lie', 'ear', 'receive', 'eve', 'miracle', 'emacs', 'cinema', 'canna', 'cascara', 'maniacal', 'manna', 'servile', 'iliac', 'revisal', 'crave', 'carnal', 'aerie', 'insincere', 'mire', 'seal', 'venial', 'less', 'recessive', 'serve', 'reel', 'case', 'larva', 'inn', 'manse', 'alive', 'iris', 'villa', 'sacral', 'clever', 'emcee', 'similar', 'vie', 'ass', 'recline', 'saline', 'learn', 'scram', 'cress', 'air', 'lean', 'mien', 'mini', 'car', 'carrel', 'sanicle', 'are', 'mile', 'missile', 'lase', 'see', 'mica', 'mimesis', 'care', 'laminar', 'salaam', 'criminal', 'liven', 'marina', 'venal', 'millenarian', 'raze', 'ail', 'salve', 'scar', 'verve', 'call', 'ace', 'millennia', 'scale', 'silane', 'lisle', 'civic', 'linear', 'melamine', 'mamma', 'larvae', 'crisis', 'sic', 'slave', 'lane', 'claim', 'incisive', 'invasive', 'assai', 'sill', 'elver', 'size', 'renal', 'sane', 'alai', 'seaman', 'seamen', 'sliver', 'virile', 'lineal', 'realm', 'alia', 'severe', 'sear', 'camellia', 'cell', 'malice', 'lever', 'lain', 'massive',
-                                    'riverine', 'clearance', 'main', 'eerie', 'larval', 'marvel', 'nemesis', 'easel', 'almanac', 'came', 'amaze', 'cream', 'airman', 'meal', 'seam', 'cease', 'leaven', 'carcass', 'mince', 'rasa', 'lice', 'messianic', 'sari', 'miser', 'liar', 'canvass', 'vise', 'liana', 'mania', 'marlin', 'lamellar', 'recess', 'mass', 'ream', 'zazen', 'snare', 'immerse', 'clan', 'eleven', 'cavern', 'aim', 'snarl', 'czar', 'alias', 'salesmen', 'mess', 'lease', 'caravan', 'save', 'scam', 'vice', 'amiss', 'rain', 'caviar', 'veal', 'visa', 'aisle', 'cassava', 'else', 'assess', 'linen', 'seminal', 'sierra', 'seven', 'airline', 'avail', 'vain', 'viz', 'cavil', 'live', 'men', 'cellar', 'maniac', 'licensee'],
-                              orientations=[['arci', 'szev', 'llae', 'msna'], ['ansm', 'eall', 'vezs', 'icra'], ['arci', 'szev', 'llae', 'msna']]
-                               )
+
+def seed_letter_boards():
+  letter_boards = [['hleo', 'aper', 'diua', 'udrt'], ['sayd', 'iunt', 'bsep', 'oris'], ['xsil', 'dend', 'peyr', 'okaa'],
+                   ['ankle', 'alai', 'kbts', 'eeas'], ['pgah', 'slai', 'eraw', 'vnoe'], ['etar', 'rsaw', 'pcvo', 'lsex'],
+                   ['mafd', 'poee', 'rdsk', 'eeap'], ['ytui', 'ehai', 'solj', 'spli'], ['voak', 'esnm', 'sahl', 'loef'],
+                   ['siox', 'cgny', 'sahl', 'itza']
+                   ]
+  letter_board_seeds = []
+  for board_letters in letter_boards:
+    letter_board_seed = Letter_Board(letters=board_letters, words=find_all_words(board_letters), orientations=find_all_rotations(board_letters))
+    letter_board_seeds.append(letter_board_seed)
+    print(find_all_rotations(board_letters))
+  print(letter_board_seeds)
+
+  db.session.add_all(letter_board_seeds)
+  db.session.commit()
+
+
+def undo_letter_boards():
+  db.session.execute('TRUNCATE letter_boards CASCADE;')
+  db.session.commit()
